+++ src/plugins/postgre-connector.js
@@ -1,7 +1,7 @@
 const Sequelize = require('sequelize')
-const sequelize = new Sequelize('dbg', 'postgres', 'metal', {
-	host: 'localhost',
-	dialect: 'postgres',
+const sequelize = new Sequelize(process.env.DB_SQL_NAME, process.env.DB_SQL_USER, process.env.DB_SQL_PASSWORD, {
+	host: process.env.DB_SQL_HOST,
+	dialect: process.env.DB_SQL_CLIENT,
 	logging: false,
 	dialectOptions: {
 		connectTimeout: 0