+++ src/server.js
@@ -1,12 +1,18 @@
 const express = require('express')
 const pjson = require('../package.json')
+const dotenv = require('dotenv')
+// const https = require('https')
+const http = require('http')
+// const fs = require('fs')
 
 const app = express()	// Application creation
 
+dotenv.config()
+
 // CORS
-// const cors = require('cors')
-// const corsOptions = { origin: 'http://localhost:3000' }
-// app.use(cors(corsOptions))
+const cors = require('cors')
+const corsOptions = { origin: 'http://localhost:3000' }
+app.use(cors(corsOptions))
 
 // Body Parser
 app.use(express.json())
@@ -22,63 +28,35 @@ global.include = function(file) {
 }
 
 // Declare routes
-require('../src/routes/routes')(app) // Declare API Routes
+include('routes/routes')(app)
 
-// Start Server
-const port = 3000
-app.listen(port, function() {
-	console.log( pjson.name + '@' + pjson.version + ' running on port ' + port)
+const httpServer = http.createServer(app)
+
+httpServer.listen(process.env.APP_PORT_HTTP, () => {
+	console.log( pjson.name + '@' + pjson.version + ' running on port ' + process.env.APP_PORT_HTTP)
 })
 
-// #44 Passer openpharma en https
-// Set port, listen for requests in HTTPS, see: Passer openpharma en https #44 (https://github.com/Bworld-Studio/openpharma/issues/44)
-
-// const https = require('https')
-// const fs = require('fs')
-
-// const port = 3443
-
+// HTTPS
 // const options = {
-// 	key: fs.readFileSync('./key.pem'),
-// 	cert: fs.readFileSync('./cert.pem'),
-// 	passphrase: 'metal01'
+// 	key: fs.readFileSync('./cert/localhost-key.pem'),
+// 	cert: fs.readFileSync('./cert/localhost.pem'),
+// 	ca: fs.readFileSync('C:\\Users\\grayf\\AppData\\Local\\mkcert\\rootCA.pem'),
 // }
 
-// https.createServer(options, app).listen(port, function() {
+// const httpsServer = https.createServer(options, app)
+
+// httpsServer.listen(process.env.APP_PORT_HTTPS, () => {
+// 	console.log( pjson.name + '@' + pjson.version + ' running on port ' + process.env.APP_PORT_HTTPS)
+// })
+
+// Start Server
+// const port = 3443
+// app.listen(port, function() {
 // 	console.log( pjson.name + '@' + pjson.version + ' running on port ' + port)
 // })
 
+// const httpsServer = https.createServer(options, app)
 
-// Fastify Server																								//
-
-// const fs = require('fs')
-// const common = require('./common/common')
-
-// let date = common.getTimeStramp()
-// let logFile = '.logs//server_'+date+'.log'
-
-// fs.writeFile(logFile)
-
-// const fastify = require('fastify')({
-// 	logger: {
-// 		level: 'info',
-// 		file: logFile // Will use pino.destination()
-// 	}
-// })
-// const pjson = require('./package.json')
-
-// // Set a GET route "/"
-// fastify.register(require('./routes/common'))
-// fastify.register(require('./routes/clients'))
-// // fastify.register(require('./routes/orders'))
-// // fastify.register(require('./routes/products'))
-// // fastify.register(require('./routes/template'))
-// // fastify.register(require('./routes/updates'))
-
-// fastify.listen(3000, function (err, address) { // eslint-disable-line no-unused-vars
-// 	if (err) {
-// 		fastify.log.error(err)
-// 		process.exit(1) // eslint-disable-line no-undef
-// 	}
-// 	fastify.log.info( pjson.name + '@' + pjson.version + ' running on port 3000')
+// httpsServer.listen(port, () => {
+// 	console.log( pjson.name + '@' + pjson.version + ' running on port ' + port)
 // })