+++ src/services/clients.py
@@ -1,2 +1,0 @@
-var = 'Openpharma-server 0.3'
-print(var)