+++ src/services/models.js
@@ -1,4 +1,5 @@
-
+/* eslint-disable no-undef */
+/* eslint-disable no-unused-vars */
 // const { Op } = require('sequelize')
 // const Client = require('../../models/Client')
 // const { v4: uuidv4 } = require('uuid')