+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 0.3.2
+  - ✅ Bump to 0.3.2
+  - ⬆️ Updates all depedencies
+  - ❌ Remove unecessary depedencies
+  - 👕 Implement "include" function to use relative paths instead of absolute in the "require" function
+  - 🚧 Splitting db queries from Api files into Services Files
+  - 🚧 Start working on Product view
+  - 🔑 Start the use of .env files
+
 ### 0.3.1
   - ✅ Bump to 0.3.1
   - ⬆️ Updates all depedencies